+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Change size detection from stty to check TTY on the output method
 * Change size detection from the environment variables to use the env method
 * Change size from the ansicon environment variable to use the env method
+* Change size from ansicon to use regexp last match to lookup pattern matches
 
 ### Fixed
 * Fix ioctl call test to stub terminal size encoding for big-endian systems