+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Fix to skip terminal size detection from readline on non-TTY output
 * Fix size detection from IO to skip loading io-console on non-TTY output
 * Fix size detection from tput to check for non-zero columns instead of lines
+* Fix JRuby and Windows system detection to check Ruby configuration only once
 
 ## [v0.8.1] - 2020-07-17